+++ handheld/src/client/gui/screens/PrerenderTilesScreen.h
@@ -65,7 +65,7 @@ public:
 			Tile::web->id, 0,
 			Item::sign->id, 0,
 		};
-		for (int i = 0; i < sizeof(items)/sizeof(int); i += 2) {
+		for (size_t i = 0; i < sizeof(items)/sizeof(int); i += 2) {
 			ItemInstance item(items[i], 1, items[i+1]);
 			addItem(&item);
 		}
